+++ src/core/linked-list/index.ts
@@ -1,0 +1,14 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+/**
+ * [[include:core/linked-list/README.md]]
+ * @packageDocumentation
+ */
+
+export { default } from 'core/linked-list/list';