+++ src/core/queue/worker/merge/index.ts
@@ -28,10 +28,10 @@ import type {
 export * from 'core/queue/worker/merge/interface';
 
 /**
- * Implementation of a worker queue data structure with support of task merging by the specified hash function
+ * Implementation of a worker queue data structure with support of task merging by a specified hash function
  *
- * @typeparam T - task element
- * @typeparam V - worker value
+ * @typeparam T - the task element
+ * @typeparam V - the worker value
  */
 export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> {
 	override readonly Tasks!: Tasks<string>;
@@ -41,24 +41,24 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 			return undefined;
 		}
 
-		const obj = this.tasksMap[this.tasks.head!];
+		const obj = this.tasksMap.get(this.tasks.head!);
 		return obj?.task;
 	}
 
 	/**
-	 * The map of registered tasks
+	 * A map of registered tasks
 	 */
-	protected tasksMap: Dictionary<Task<T, V>> = Object.createDict();
+	protected tasksMap: Map<string, Task<T, V>> = new Map();
 
 	/**
-	 * Function to calculate a task hash
+	 * A function to calculate task hashes
 	 */
 	protected readonly hashFn: HashFn<T>;
 
 	/**
 	 * @override
 	 * @param worker
-	 * @param [opts]
+	 * @param [opts] - additional options
 	 */
 	constructor(worker: QueueWorker<T, V>, opts: WorkerQueueOptions<T>) {
 		super(worker, opts);
@@ -70,7 +70,7 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 			hash = this.hashFn(task);
 
 		let
-			taskObj = this.tasksMap[hash];
+			taskObj = this.tasksMap.get(hash);
 
 		if (taskObj == null) {
 			let
@@ -81,7 +81,7 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 			});
 
 			taskObj = {task, promise, resolve};
-			this.tasksMap[hash] = taskObj;
+			this.tasksMap.set(hash, taskObj);
 			this.tasks.push(hash);
 		}
 
@@ -97,7 +97,7 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 		const
 			{head} = this;
 
-		delete this.tasksMap[this.tasks.head!];
+		this.tasksMap.delete(this.tasks.head!);
 		this.tasks.shift();
 
 		return head;
@@ -106,8 +106,32 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 	override clear(): void {
 		if (this.length > 0) {
 			super.clear();
-			this.tasksMap = Object.createDict();
+			this.tasksMap = new Map<string, Task<T, V>>();
 		}
+	}
+
+	override clone(): MergeWorkerQueue<T, V> {
+		const
+			newQueue = new MergeWorkerQueue<T, V>(this.worker, {});
+
+		Object.assign(newQueue, this);
+		newQueue.tasksMap = new Map(this.tasksMap);
+
+		if (this.tasks.clone != null) {
+			newQueue.tasks = this.tasks.clone();
+
+		} else {
+			const
+				tasks = this.createTasks();
+
+			for (const task of this.tasks) {
+				tasks.push(task);
+			}
+
+			newQueue.tasks = tasks;
+		}
+
+		return newQueue;
 	}
 
 	protected override perform(): void {
@@ -124,9 +148,9 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 		}
 
 		const
-			taskObj = this.tasksMap[hash];
+			taskObj = this.tasksMap.get(hash);
 
-		if (!taskObj) {
+		if (taskObj == null) {
 			return;
 		}
 
@@ -135,7 +159,7 @@ export default class MergeWorkerQueue<T, V = unknown> extends WorkerQueue<T, V> 
 			{task, promise, resolve} = taskObj;
 
 		const cb = () => {
-			delete this.tasksMap[hash];
+			this.tasksMap.delete(hash);
 			return this.deferPerform();
 		};